Refactor complete and compiling
