+++ Cargo.toml
@@ -62,6 +62,7 @@ public-credentials = {path = "pallets/public-credentials", default-features = fa
 # Internal support (with default disabled)
 dip-support = {path = "crates/dip-support", default-features = false}
 kilt-asset-dids = {path = "crates/assets", default-features = false}
+kilt-dip-support = {path = "crates/kilt-dip-support", default-features = false}
 kilt-support = {path = "support", default-features = false}
 runtime-common = {path = "runtimes/common", default-features = false}
