+++ crates/dip-support/Cargo.toml
@@ -17,13 +17,11 @@ scale-info = {workspace = true, features = ["derive"]}
 
 # Substrate dependencies
 frame-support.workspace = true
-sp-std.workspace = true
 
 [features]
 default = ["std"]
 std = [
   "parity-scale-codec/std",
   "scale-info/std",
-  "frame-support/std",
-  "sp-std/std"
+  "frame-support/std"
 ]
