+++ crates/kilt-dip-support/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+authors.workspace = true
+description = "Support types, traits, and functions for the KILT Decentralized Identity Provider (DIP) functionality as implemented by the KILT blockchain."
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+name = "kilt-dip-support"
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+# Internal dependencies
+did.workspace = true
+pallet-dip-consumer.workspace = true
+
+# Parity dependencies
+parity-scale-codec = {workspace = true, features = ["derive"]}
+scale-info = {workspace = true, features = ["derive"]}
+
+# Substrate dependencies
+frame-support.workspace = true
+sp-runtime.workspace = true
+sp-core.workspace = true
+sp-trie.workspace = true
+sp-std.workspace = true
+
+[features]
+default = ["std"]
+std = [
+  "did/std",
+  "pallet-dip-consumer/std",
+  "parity-scale-codec/std",
+  "scale-info/std",
+  "frame-support/std",
+  "sp-runtime/std",
+  "sp-core/std",
+  "sp-trie/std",
+  "sp-std/std"
+]
+runtime-benchmarks = []
