+++ crates/kilt-dip-support/src/lib.rs
@@ -16,12 +16,10 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use frame_support::RuntimeDebug;
-use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
-use scale_info::TypeInfo;
+// TODO: Crate documentation
 
-#[derive(Clone, Eq, PartialEq, Encode, Decode, TypeInfo, MaxEncodedLen, RuntimeDebug)]
-pub enum IdentityProofAction<Identifier, Proof, Details = ()> {
-	Updated(Identifier, Proof, Details),
-	Deleted(Identifier),
-}
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod did;
+pub mod merkle;
+pub mod traits;
