+++ runtimes/common/src/dip/did.rs
@@ -16,12 +16,27 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use frame_support::RuntimeDebug;
-use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
-use scale_info::TypeInfo;
+use did::did_details::DidDetails;
+use pallet_dip_provider::traits::IdentityProvider;
+use sp_std::marker::PhantomData;
 
-#[derive(Clone, Eq, PartialEq, Encode, Decode, TypeInfo, MaxEncodedLen, RuntimeDebug)]
-pub enum IdentityProofAction<Identifier, Proof, Details = ()> {
-	Updated(Identifier, Proof, Details),
-	Deleted(Identifier),
+pub struct DidIdentityProvider<T>(PhantomData<T>);
+
+impl<T> IdentityProvider<T::DidIdentifier, DidDetails<T>, ()> for DidIdentityProvider<T>
+where
+	T: did::Config,
+{
+	// TODO: Proper error handling
+	type Error = ();
+
+	fn retrieve(identifier: &T::DidIdentifier) -> Result<Option<(DidDetails<T>, ())>, Self::Error> {
+		match (
+			did::Pallet::<T>::get_did(identifier),
+			did::Pallet::<T>::get_deleted_did(identifier),
+		) {
+			(Some(details), _) => Ok(Some((details, ()))),
+			(_, Some(_)) => Ok(None),
+			_ => Err(()),
+		}
+	}
 }
